+++ pcigale/session/configuration.py
@@ -93,18 +93,42 @@ class Configuration(object):
 
         self.config['sed_modules'] = []
         self.config.comments['sed_modules'] = ([""] +
-            ["Order of the modules use for SED creation. Available modules:"] +
-            ["SFH: sfh2exp, sfhdelayed, sfhdelayedbq, sfhfromfile, "
-             "sfhperiodic"] +
-            ["SSP: bc03, m2005"] +
-            ["Nebular emission: nebular"] +
-            ["Dust attenuation: dustatt_calzleit, dustatt_powerlaw, "
-             "dustatt_2powerlaws"] +
-            ["Dust emission: casey2012, dale2014, dl2007, dl2014, themis"] +
-            ["AGN: dale2014, fritz2006"] +
-            ["Radio: radio"] +
-            ["Restframe parameters: restframe_parameters"] +
-            ["Redshift: redshifting (mandatory!)"])
+            ["Avaiable modules to compute the models. The order must be kept."
+            ] +
+            ["SFH:"] +
+            ["* sfh2exp (double exponential)"] +
+            ["* sfhdelayed (delayed SFH with optional exponential burst)"] +
+            ["* sfhdelayedbq (delayed SFH with optional constant burst/quench)"
+            ] +
+            ["* sfhfromfile (abitrary SFH read from an input file)"] +
+            ["* sfhperiodic (periodic SFH, exponential, rectangle or delayed"
+             ")"] +
+            ["SSP:"] +
+            ["* bc03 (Bruzual and Charlot 2003)"] +
+            ["* m2005 (Maraston 2005)"] +
+            ["Nebular emission:"] +
+            ["* nebular (continuum and line nebular emission)"] +
+            ["Dust attenuation:"] +
+            ["* dustatt_modified_CF00 (modified Charlot & Fall 2000 "
+             "attenuation law)"] +
+            ["* dustatt_modified_starburst (modified starburst attenuaton law)"
+            ] +
+            ["Dust emission:"] +
+            ["* casey2012 (Casey 2012 dust emission models)"] +
+            ["* dale2014 (Dale et al. 2014 dust emission templates)"] +
+            ["* dl2007 (Draine & Li 2007 dust emission models)"] +
+            ["* dl2014 (Draine et al. 2014 update of the previous models)"] +
+            ["* themis (Themis dust emission models from Jones et al. 2017)"] +
+            ["AGN:"] +
+            ["* fritz2006 (AGN models from Fritz et al. 2006)"] +
+            ["Radio:"] +
+            ["* radio (synchrotron emission)"] +
+            ["Restframe parameters:"] +
+            ["* restframe_parameters (UV slope, IRX-beta, D4000, EW, etc.)"] +
+            ["Redshift+IGM:"] +
+            ["* redshifting (mandatory, also includes the IGM from Meiksin "
+             "2006)"]
+        )
         self.spec['sed_modules'] = "cigale_string_list()"
 
         self.config['analysis_method'] = ""